+++ src/features/index.ts
@@ -1,3 +1,1 @@
-export { ActivityTab } from "./activity/ActivityTab";
 export { StoreTab } from "./store/StoreTab";
-export { SystemTab, appendSessionLog } from "./system/SystemTab";